requirements.txt
addict==2.4.0
anyio==4.10.0
certifi==2025.8.3
charset-normalizer==3.4.3
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
lark-oapi==1.4.21
pycryptodome==3.23.0
python-dotenv==1.1.1
requests==2.32.5
requests-toolbelt==1.0.0
sniffio==1.3.1
sqlalchemy==2.0.43
typing-extensions==4.14.1
urllib3==2.5.0
websockets==15.0.1
